+++ src/map/packets/bazaar_item.h
@@ -40,7 +40,7 @@ class CBazaarItemPacket: public CBasicPacket
 {
 public:
 
-	CBazaarItemPacket(CItem* PItem, uint8 SlotID, uint8 Tax);
+	CBazaarItemPacket(CItem* PItem, uint8 SlotID, uint16 Tax);
 };
 
 #endif
